# docs/swagger.yaml
definitions:
...
  entity.Olahraga:
    properties:
      deskripsi:
        type: string
      id:
        type: integer
      nama:
        type: string
    type: object
...
  entity.User:
    properties:
      email:
        type: string
      id:
        type: integer
      id_imt:
        type: integer
      id_kesehatan:
        type: integer
      name:
        type: string
      password:
        type: string
    type: object
host: localhost:8888
info:
  contact: {}
  description: This is Fits Health API
  title: Fits Health API Documentation
  version: "2.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Login User
      parameters:
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.User'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/entity.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.User'
      summary: Login User
      tags:
      - User
  /olahraga:
    get:
      consumes:
      - application/json
      description: User can get information about all workout that can do it in home
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Olahraga'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Olahraga'
      security:
      - JWT: []
      summary: Get All Workout Information
      tags:
      - Olahraga
    post:
      consumes:
      - application/json
      description: Developer can create workout for User
      parameters:
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.Olahraga'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Olahraga'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Olahraga'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Olahraga'
      security:
      - JWT: []
      summary: Create Workout
      tags:
      - Olahraga
  /olahraga/{id}:
    delete:
      consumes:
      - application/json
      description: Developer can Delete Workout Information if workout information
        is useless
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Olahraga'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Olahraga'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Olahraga'
      security:
      - JWT: []
      summary: Delete Workout
      tags:
      - Olahraga
    get:
      consumes:
      - application/json
      description: User can get Workout Information by ID
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Olahraga'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Olahraga'
      security:
      - JWT: []
      summary: Get Workout Information by ID
      tags:
      - Olahraga
    put:
      consumes:
      - application/json
      description: Developer can update about workout information
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Olahraga'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.Olahraga'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.Olahraga'
      security:
      - JWT: []
      summary: Update Olahraga
      tags:
      - Olahraga
...
  /register:
    post:
      consumes:
      - application/json
      description: People can Register as a User
      parameters:
      - description: required
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.User'
      summary: Register User
      tags:
      - User
...
  /users:
    get:
      consumes:
      - application/json
      description: Get ALl User
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.User'
      security:
      - JWT: []
      summary: Get All User
      tags:
      - User
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: User can Delete their own account
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.User'
      security:
      - JWT: []
      summary: Delete User
      tags:
      - User
    get:
      consumes:
      - application/json
      description: User Can Get User Information By Id
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.User'
      security:
      - JWT: []
      summary: Get User By Id
      tags:
      - User
    put:
      consumes:
      - application/json
      description: User can Update their status or information
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entity.User'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entity.User'
      security:
      - JWT: []
      summary: Update User
      tags:
      - User
schemes:
- http
- https
securityDefinitions:
  JWT:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"